feat(ui): introduce a basic protocol view page
